requirements.txt
torch~=1.5.1
torchvision~=0.6.1
opencv-python~=4.3.0.36
numpy~=1.19.0
pandas~=1.3.0
Pillow~=7.2.0
PyYaml
setuptools~=57.0.0
